#  core: rework and optimize SIMD implementation of dotProd