… modules/core/CMakeLists.txt
… modules/core/include/opencv2/core/hal/intrin_cpp.hpp
… modules/core/include/opencv2/core/hal/intrin_msa.hpp
… modules/core/include/opencv2/core/hal/intrin_neon.hpp
… modules/core/include/opencv2/core/hal/intrin_sse.hpp
… modules/core/include/opencv2/core/hal/intrin_vsx.hpp
… modules/core/include/opencv2/core/hal/intrin_wasm.hpp
… modules/core/test/test_intrin_utils.hpp