relates #15506
merge with opencv_extra opencv/opencv_extra#674

### This pullrequest changes
   - add new universal intrinsics v_dotprod[int32], v_dotprod_expand[u&int8, u&int16, int32], v_cvt_f64(int64)
  - add a boolean param for all v_dotprod&_expand intrinsics that change the behavior of addition order between
    pairs in some platforms in order to reach the maximum optimization when the sum among all lanes is what only matters
  - fix clang build on ppc64le
  - support wide universal intrinsics for dotProd_32s
  - remove raw SIMD and activate universal intrinsics for dotProd_8
  - implement SIMD optimization for dotProd_s16&u16
  - extend performance test data types of dot prod
  - fix GCC VSX workaround of vec_mule and vec_mulo (in little-endian must be swapped)
  - optimize v_mul_expand(int32) on VSX

### TODO
- [x] performance tests
- [x] ppc64le clang coverage 
- [x]  cover msa & wasm
- [x]  update perf data

````
force_builders=Custom,armv7,armv8,linux
buildworker:Custom=linux-1
docker_image:Custom=mips64el
pw_compilers=gcc-4.9, power9_gcc-5, power9_gcc-6, power9_gcc-7, clang-4, clang-5, clang-6
pw_with_opencv_extra=branch:issue15506
# https://ocv-power.imavr.com/#/builders/3/builds/84
# imgproc test crash on clang 4, 5 . I disable it for now
pw_disable_tests=rgbd, shape, imgproc
````
## Performance tests
#### Args used with opencv_perf_core
```
--gtest_filter=*dot* --perf_threads=1 --perf_min_samples=1000 --gtest_output=xml:[path]
```

### X86
<details>
<summary>CPU</summary>

```
Architecture:          x86_64
CPU op-mode(s):        32-bit, 64-bit
Byte Order:            Little Endian
CPU(s):                8
On-line CPU(s) list:   0-7
Thread(s) per core:    2
Core(s) per socket:    4
Socket(s):             1
NUMA node(s):          1
Vendor ID:             GenuineIntel
CPU family:            6
Model:                 142
Model name:            Intel(R) Core(TM) i7-8550U CPU @ 1.80GHz
Stepping:              10
CPU MHz:               818.725
CPU max MHz:           4000.0000
CPU min MHz:           400.0000
BogoMIPS:              3984.00
Virtualization:        VT-x
L1d cache:             32K
L1i cache:             32K
L2 cache:              256K
L3 cache:              8192K
NUMA node0 CPU(s):     0-7
Flags:                 fpu vme de pse tsc msr pae mce cx8 apic sep mtrr pge mca cmov pat pse36 clflush dts acpi mmx fxsr sse sse2 ss ht tm pbe syscall nx pdpe1gb rdtscp lm constant_tsc art arch_perfmon pebs bts rep_good nopl xtopology nonstop_tsc aperfmperf pni pclmulqdq dtes64 monitor ds_cpl vmx est tm2 ssse3 sdbg fma cx16 xtpr pdcm pcid sse4_1 sse4_2 x2apic movbe popcnt tsc_deadline_timer aes xsave avx f16c rdrand lahf_lm abm 3dnowprefetch epb invpcid_single ssbd ibrs ibpb stibp kaiser tpr_shadow vnmi flexpriority ept vpid fsgsbase tsc_adjust bmi1 avx2 smep bmi2 erms invpcid mpx rdseed adx smap clflushopt intel_pt xsaveopt xsavec xgetbv1 xsaves dtherm ida arat pln pts hwp hwp_notify hwp_act_window hwp_epp flush_l1d
```
</details>
<details>
<summary>OS</summary>

```
Linux seiko-pc 4.19.0-5-amd64 #1 SMP Debian 4.19.37-5 (2019-06-19) x86_64 GNU/Linux
gcc version 8.3.0 (Debian 8.3.0-2)
Debian GNU/Linux 10 (buster)
```
</details>

#### Benchmark

<details>
 <summary>BASELINE AVX2 Geometric mean (ms)</summary>

|Name of Test|Before|After|After vs Before (x-factor)|
|---|:-:|:-:|:-:|
|dot::MatType_Length::(8UC1, 32)|0.000|0.000|1.16|
|dot::MatType_Length::(8UC1, 64)|0.000|0.000|1.23|
|dot::MatType_Length::(8UC1, 128)|0.001|0.001|1.26|
|dot::MatType_Length::(8UC1, 256)|0.003|0.003|1.24|
|dot::MatType_Length::(8UC1, 512)|0.014|0.012|1.21|
|dot::MatType_Length::(8UC1, 1024)|0.054|0.046|1.19|
|dot::MatType_Length::(8SC1, 32)|0.000|0.000|1.15|
|dot::MatType_Length::(8SC1, 64)|0.000|0.000|1.28|
|dot::MatType_Length::(8SC1, 128)|0.001|0.001|1.34|
|dot::MatType_Length::(8SC1, 256)|0.003|0.003|1.33|
|dot::MatType_Length::(8SC1, 512)|0.014|0.011|1.18|
|dot::MatType_Length::(8SC1, 1024)|0.051|0.045|1.13|
|dot::MatType_Length::(16UC1, 32)|0.001|0.000|4.77|
|dot::MatType_Length::(16UC1, 64)|0.003|0.000|6.30|
|dot::MatType_Length::(16UC1, 128)|0.010|0.002|6.20|
|dot::MatType_Length::(16UC1, 256)|0.041|0.007|6.01|
|dot::MatType_Length::(16UC1, 512)|0.165|0.027|6.05|
|dot::MatType_Length::(16UC1, 1024)|0.665|0.112|5.94|
|dot::MatType_Length::(16SC1, 32)|0.001|0.000|6.08|
|dot::MatType_Length::(16SC1, 64)|0.003|0.000|8.86|
|dot::MatType_Length::(16SC1, 128)|0.010|0.001|9.09|
|dot::MatType_Length::(16SC1, 256)|0.041|0.005|8.56|
|dot::MatType_Length::(16SC1, 512)|0.166|0.021|8.02|
|dot::MatType_Length::(16SC1, 1024)|0.668|0.085|7.85|
|dot::MatType_Length::(32SC1, 32)|0.000|0.000|1.35|
|dot::MatType_Length::(32SC1, 64)|0.001|0.001|1.35|
|dot::MatType_Length::(32SC1, 128)|0.004|0.003|1.40|
|dot::MatType_Length::(32SC1, 256)|0.017|0.013|1.34|
|dot::MatType_Length::(32SC1, 512)|0.069|0.051|1.35|
|dot::MatType_Length::(32SC1, 1024)|0.335|0.264|1.27|
|dot::MatType_Length::(32FC1, 32)|0.000|0.000|0.98|
|dot::MatType_Length::(32FC1, 64)|0.000|0.000|0.97|
|dot::MatType_Length::(32FC1, 128)|0.002|0.002|1.01|
|dot::MatType_Length::(32FC1, 256)|0.010|0.010|1.02|
|dot::MatType_Length::(32FC1, 512)|0.039|0.038|1.00|
|dot::MatType_Length::(32FC1, 1024)|0.247|0.249|0.99|

</details>


<details>
 <summary>BASELINE SSE4_1 Geometric mean (ms)</summary>

|Name of Test|Before|After|After vs Before (x-factor)|
|---|:-:|:-:|:-:|
|dot::MatType_Length::(8UC1, 32)|0.000|0.000|0.99|
|dot::MatType_Length::(8UC1, 64)|0.000|0.000|0.98|
|dot::MatType_Length::(8UC1, 128)|0.001|0.001|0.97|
|dot::MatType_Length::(8UC1, 256)|0.005|0.005|0.97|
|dot::MatType_Length::(8UC1, 512)|0.020|0.020|1.00|
|dot::MatType_Length::(8UC1, 1024)|0.079|0.076|1.04|
|dot::MatType_Length::(8SC1, 32)|0.000|0.000|1.00|
|dot::MatType_Length::(8SC1, 64)|0.000|0.000|1.00|
|dot::MatType_Length::(8SC1, 128)|0.001|0.001|1.00|
|dot::MatType_Length::(8SC1, 256)|0.005|0.005|0.99|
|dot::MatType_Length::(8SC1, 512)|0.018|0.019|0.98|
|dot::MatType_Length::(8SC1, 1024)|0.073|0.073|1.00|
|dot::MatType_Length::(16UC1, 32)|0.001|0.000|3.05|
|dot::MatType_Length::(16UC1, 64)|0.003|0.001|3.43|
|dot::MatType_Length::(16UC1, 128)|0.011|0.003|3.54|
|dot::MatType_Length::(16UC1, 256)|0.044|0.012|3.55|
|dot::MatType_Length::(16UC1, 512)|0.177|0.050|3.53|
|dot::MatType_Length::(16UC1, 1024)|0.707|0.199|3.55|
|dot::MatType_Length::(16SC1, 32)|0.001|0.000|5.09|
|dot::MatType_Length::(16SC1, 64)|0.003|0.000|6.49|
|dot::MatType_Length::(16SC1, 128)|0.011|0.002|6.97|
|dot::MatType_Length::(16SC1, 256)|0.044|0.006|6.92|
|dot::MatType_Length::(16SC1, 512)|0.177|0.026|6.84|
|dot::MatType_Length::(16SC1, 1024)|0.708|0.103|6.85|
|dot::MatType_Length::(32SC1, 32)|0.000|0.000|1.02|
|dot::MatType_Length::(32SC1, 64)|0.001|0.001|1.06|
|dot::MatType_Length::(32SC1, 128)|0.005|0.005|1.01|
|dot::MatType_Length::(32SC1, 256)|0.019|0.018|1.05|
|dot::MatType_Length::(32SC1, 512)|0.078|0.074|1.06|
|dot::MatType_Length::(32SC1, 1024)|0.361|0.346|1.04|
|dot::MatType_Length::(32FC1, 32)|0.000|0.000|1.01|
|dot::MatType_Length::(32FC1, 64)|0.000|0.000|0.95|
|dot::MatType_Length::(32FC1, 128)|0.002|0.002|0.99|
|dot::MatType_Length::(32FC1, 256)|0.010|0.010|1.00|
|dot::MatType_Length::(32FC1, 512)|0.038|0.039|0.99|
|dot::MatType_Length::(32FC1, 1024)|0.261|0.236|1.11|


</details>

<details>
 <summary>BASELINE SSE3 Geometric mean (ms)</summary>

|Name of Test|Before|After|After vs Before (x-factor)|
|---|:-:|:-:|:-:|
|dot::MatType_Length::(8UC1, 32)|0.000|0.000|0.99|
|dot::MatType_Length::(8UC1, 64)|0.000|0.000|0.99|
|dot::MatType_Length::(8UC1, 128)|0.001|0.001|1.00|
|dot::MatType_Length::(8UC1, 256)|0.005|0.005|1.00|
|dot::MatType_Length::(8UC1, 512)|0.018|0.019|0.99|
|dot::MatType_Length::(8UC1, 1024)|0.072|0.073|0.97|
|dot::MatType_Length::(8SC1, 32)|0.000|0.000|1.10|
|dot::MatType_Length::(8SC1, 64)|0.000|0.000|1.10|
|dot::MatType_Length::(8SC1, 128)|0.001|0.001|1.12|
|dot::MatType_Length::(8SC1, 256)|0.006|0.005|1.11|
|dot::MatType_Length::(8SC1, 512)|0.023|0.021|1.09|
|dot::MatType_Length::(8SC1, 1024)|0.092|0.085|1.08|
|dot::MatType_Length::(16UC1, 32)|0.001|0.000|3.05|
|dot::MatType_Length::(16UC1, 64)|0.003|0.001|3.42|
|dot::MatType_Length::(16UC1, 128)|0.011|0.003|3.52|
|dot::MatType_Length::(16UC1, 256)|0.044|0.013|3.51|
|dot::MatType_Length::(16UC1, 512)|0.177|0.050|3.54|
|dot::MatType_Length::(16UC1, 1024)|0.707|0.202|3.51|
|dot::MatType_Length::(16SC1, 32)|0.001|0.000|5.16|
|dot::MatType_Length::(16SC1, 64)|0.003|0.000|6.50|
|dot::MatType_Length::(16SC1, 128)|0.011|0.002|6.93|
|dot::MatType_Length::(16SC1, 256)|0.044|0.007|6.74|
|dot::MatType_Length::(16SC1, 512)|0.177|0.026|6.75|
|dot::MatType_Length::(16SC1, 1024)|0.708|0.108|6.58|
|dot::MatType_Length::(32SC1, 32)|0.000|0.000|1.03|
|dot::MatType_Length::(32SC1, 64)|0.001|0.001|1.06|
|dot::MatType_Length::(32SC1, 128)|0.005|0.005|1.04|
|dot::MatType_Length::(32SC1, 256)|0.020|0.019|1.02|
|dot::MatType_Length::(32SC1, 512)|0.078|0.076|1.03|
|dot::MatType_Length::(32SC1, 1024)|0.367|0.364|1.01|
|dot::MatType_Length::(32FC1, 32)|0.000|0.000|1.00|
|dot::MatType_Length::(32FC1, 64)|0.000|0.000|0.97|
|dot::MatType_Length::(32FC1, 128)|0.002|0.002|1.00|
|dot::MatType_Length::(32FC1, 256)|0.010|0.010|1.00|
|dot::MatType_Length::(32FC1, 512)|0.042|0.042|0.99|
|dot::MatType_Length::(32FC1, 1024)|0.243|0.238|1.02|
</details>

-------------

### PPC64LE
<details>
<summary>CPU</summary>

```
Architecture:        ppc64le
Byte Order:          Little Endian
CPU(s):              8
On-line CPU(s) list: 0-7
Thread(s) per core:  1
Core(s) per socket:  1
Socket(s):           8
NUMA node(s):        1
Model:               2.2 (pvr 004e 1202)
Model name:          POWER9 (architected), altivec supported
L1d cache:           32K
L1i cache:           32K
NUMA node0 CPU(s):   0-7
```
```
cat /proc/cpuinfo
processor	: 0
cpu		: POWER9 (architected), altivec supported
clock		: 2200.000000MHz
revision	: 2.2 (pvr 004e 1202)
timebase	: 512000000
platform	: pSeries
model		: IBM pSeries (emulated by qemu)
machine		: CHRP IBM pSeries (emulated by qemu)
MMU		: Radix

```
</details>
<details>
<summary>OS</summary>

```
Linux 8b2db3b0dfac 4.19.0-2-powerpc64le #1 SMP Debian 4.19.16-1 (2019-01-17) ppc64le ppc64le ppc64le GNU/Linux
gcc version 8.3.0 (Ubuntu 8.3.0-6ubuntu1~18.10.1)
Ubuntu 18.10 (cosmic)
```
</details>

#### Benchmark

<details>
 <summary>BASELINE VSX Geometric mean (ms)</summary>

|Name of Test|Before|After|After vs Before (x-factor)|
|---|:-:|:-:|:-:|
|dot::MatType_Length::(8UC1, 32)|0.000|0.000|1.30|
|dot::MatType_Length::(8UC1, 64)|0.001|0.001|1.63|
|dot::MatType_Length::(8UC1, 128)|0.003|0.002|1.76|
|dot::MatType_Length::(8UC1, 256)|0.012|0.006|1.84|
|dot::MatType_Length::(8UC1, 512)|0.046|0.025|1.84|
|dot::MatType_Length::(8UC1, 1024)|0.182|0.098|1.86|
|dot::MatType_Length::(8SC1, 32)|0.000|0.000|1.04|
|dot::MatType_Length::(8SC1, 64)|0.001|0.001|1.05|
|dot::MatType_Length::(8SC1, 128)|0.003|0.003|1.05|
|dot::MatType_Length::(8SC1, 256)|0.012|0.011|1.05|
|dot::MatType_Length::(8SC1, 512)|0.047|0.044|1.07|
|dot::MatType_Length::(8SC1, 1024)|0.193|0.174|1.11|
|dot::MatType_Length::(16UC1, 32)|0.001|0.001|1.84|
|dot::MatType_Length::(16UC1, 64)|0.003|0.002|2.00|
|dot::MatType_Length::(16UC1, 128)|0.012|0.006|2.03|
|dot::MatType_Length::(16UC1, 256)|0.048|0.023|2.05|
|dot::MatType_Length::(16UC1, 512)|0.193|0.093|2.07|
|dot::MatType_Length::(16UC1, 1024)|0.770|0.372|2.07|
|dot::MatType_Length::(16SC1, 32)|0.001|0.000|2.73|
|dot::MatType_Length::(16SC1, 64)|0.004|0.001|3.26|
|dot::MatType_Length::(16SC1, 128)|0.016|0.005|3.36|
|dot::MatType_Length::(16SC1, 256)|0.063|0.018|3.41|
|dot::MatType_Length::(16SC1, 512)|0.250|0.074|3.41|
|dot::MatType_Length::(16SC1, 1024)|1.001|0.293|3.41|
|dot::MatType_Length::(32SC1, 32)|0.001|0.001|1.07|
|dot::MatType_Length::(32SC1, 64)|0.003|0.002|1.12|
|dot::MatType_Length::(32SC1, 128)|0.010|0.009|1.13|
|dot::MatType_Length::(32SC1, 256)|0.040|0.035|1.13|
|dot::MatType_Length::(32SC1, 512)|0.159|0.140|1.14|
|dot::MatType_Length::(32SC1, 1024)|0.634|0.560|1.13|
|dot::MatType_Length::(32FC1, 32)|0.000|0.000|0.95|
|dot::MatType_Length::(32FC1, 64)|0.001|0.001|1.00|
|dot::MatType_Length::(32FC1, 128)|0.004|0.004|1.00|
|dot::MatType_Length::(32FC1, 256)|0.016|0.016|1.00|
|dot::MatType_Length::(32FC1, 512)|0.062|0.063|1.00|
|dot::MatType_Length::(32FC1, 1024)|0.249|0.254|0.98|
</details>

---------------

### ARM

<details>
<summary>CPU</summary>

```
Architecture:        aarch64
Byte Order:          Little Endian
CPU(s):              8
On-line CPU(s) list: 0-7
Thread(s) per core:  1
Core(s) per socket:  4
Socket(s):           2
Vendor ID:           ARM
Model:               4
Model name:          Cortex-A53
Stepping:            r0p4
CPU max MHz:         2314.0000
CPU min MHz:         403.0000
BogoMIPS:            52.00
Flags:               fp asimd evtstrm aes pmull sha1 sha2 crc32 cpuid

```
</details>
<details>
<summary>OS</summary>

```
Linux localhost 4.14.62-16641116 #1 SMP PREEMPT Fri Aug 23 14:14:12 KST 2019 aarch64 GNU/Linux
gcc version 8.3.0 (Debian 8.3.0-6)
Debian GNU/Linux 10 (buster)
```
</details>

#### Benchmark

<details>
 <summary>BASELINE NEON Geometric mean (ms)</summary>

|Name of Test|Before|After|After vs Before (x-factor)|
|---|:-:|:-:|:-:|
|dot::MatType_Length::(8UC1, 32)|0.004|0.003|1.38|
|dot::MatType_Length::(8UC1, 64)|0.007|0.005|1.46|
|dot::MatType_Length::(8UC1, 128)|0.024|0.013|1.77|
|dot::MatType_Length::(8UC1, 256)|0.095|0.044|2.17|
|dot::MatType_Length::(8UC1, 512)|0.304|0.154|1.97|
|dot::MatType_Length::(8UC1, 1024)|0.770|0.444|1.73|
|dot::MatType_Length::(8SC1, 32)|0.002|0.002|1.20|
|dot::MatType_Length::(8SC1, 64)|0.004|0.002|1.77|
|dot::MatType_Length::(8SC1, 128)|0.013|0.006|2.11|
|dot::MatType_Length::(8SC1, 256)|0.047|0.020|2.37|
|dot::MatType_Length::(8SC1, 512)|0.182|0.073|2.49|
|dot::MatType_Length::(8SC1, 1024)|0.737|0.306|2.41|
|dot::MatType_Length::(16UC1, 32)|0.003|0.002|1.46|
|dot::MatType_Length::(16UC1, 64)|0.008|0.004|1.86|
|dot::MatType_Length::(16UC1, 128)|0.029|0.013|2.16|
|dot::MatType_Length::(16UC1, 256)|0.111|0.048|2.32|
|dot::MatType_Length::(16UC1, 512)|0.444|0.200|2.22|
|dot::MatType_Length::(16UC1, 1024)|1.783|0.796|2.24|
|dot::MatType_Length::(16SC1, 32)|0.003|0.002|1.60|
|dot::MatType_Length::(16SC1, 64)|0.008|0.004|2.25|
|dot::MatType_Length::(16SC1, 128)|0.029|0.011|2.67|
|dot::MatType_Length::(16SC1, 256)|0.111|0.038|2.93|
|dot::MatType_Length::(16SC1, 512)|0.444|0.152|2.91|
|dot::MatType_Length::(16SC1, 1024)|1.783|0.615|2.90|
|dot::MatType_Length::(32SC1, 32)|0.003|0.002|1.31|
|dot::MatType_Length::(32SC1, 64)|0.008|0.005|1.53|
|dot::MatType_Length::(32SC1, 128)|0.028|0.017|1.65|
|dot::MatType_Length::(32SC1, 256)|0.107|0.063|1.71|
|dot::MatType_Length::(32SC1, 512)|0.426|0.273|1.56|
|dot::MatType_Length::(32SC1, 1024)|1.664|1.071|1.55|
|dot::MatType_Length::(32FC1, 32)|0.002|0.002|0.98|
|dot::MatType_Length::(32FC1, 64)|0.004|0.004|0.98|
|dot::MatType_Length::(32FC1, 128)|0.012|0.012|1.00|
|dot::MatType_Length::(32FC1, 256)|0.042|0.043|0.96|
|dot::MatType_Length::(32FC1, 512)|0.232|0.240|0.97|
|dot::MatType_Length::(32FC1, 1024)|0.935|0.938|1.00|

</details>