@terfendail 2019-09-13
IMHO it make sense to extend v_dotprod to v_uint32 as well